Fixed some clang warnings

+++ example/oglplus/023_waves.cpp
@@ -240,7 +240,7 @@ public:
 			attr.Enable();
 		}
 
-		Uniform<Vec3f>(prog, "LightPosition").Set(-100.0, 100.0, 20.0);
+		Uniform<Vec3f>(prog, "LightPosition").Set(-100, 100, 20);
 
 		Uniform<Vec3f> wave_directions(prog, "WaveDirections");
 		Uniform<Vec3f> wave_dimensions(prog, "WaveDimensions");
@@ -314,14 +314,14 @@ public:
 
 		auto camera = CamMatrixf::Orbiting(
 			Vec3f(0.0f, 2.0f, 0.0f),
-			17.0f - CosineWave(time / 31.0f) * 10.0f,
+			GLfloat(17.0 - CosineWave(time / 31.0) * 10.0),
 			FullCircles(time / 43.0f),
 			Degrees(45 - SineWave(time / 29.0f) * 35)
 		);
 		camera_matrix.Set(camera);
 		camera_position.Set(camera.Position());
 
-		anim_time.Set(time);
+		anim_time.Set(GLfloat(time));
 
 		plane_instr.Draw(plane_indices);
 	}
